chore: arrumando pendencias
