+++ src/components/homeDesktop/homeComponent.css
@@ -1,0 +1,105 @@
+    body {
+    background-color: var(--backgrounddesktop-color);
+    }
+    
+    .logotipo {
+    display: block;
+    position: fixed;
+    right: 0;
+    top: 25%;
+    left: 65%;
+    transform: translateY(-50%);
+    animation: move-up-down 3s ease-in-out infinite;
+    }
+    
+    .imagem {
+    display: block;
+    position: fixed;
+    right: 0;
+    top: 50%;
+    left: 20%;
+    transform: translateY(-50%);
+    }
+    
+    @keyframes move-up-down {
+    0% {
+    transform: translateY(0);
+    }
+    50% {
+    transform: translateY(-25px);
+    }
+    100% {
+    transform: translateY(0);
+    }
+    }
+    
+    h1 {
+    display: none;
+    }
+    
+    .this_button {
+    display: none;
+    }
+    
+    @media screen and (max-width: 1000px) {
+    h1 {
+    display: block;
+    }
+    
+    .imagem {
+    display: none;
+    }
+    
+    .soft {
+    color: var(--darkbtn-color);
+    }
+    
+    h1 span {
+    color: #408d86;
+    letter-spacing: 0.2rem;
+    }
+    
+    body {
+    background-color: var(--backgroundmobile-color);
+    }
+    
+    .logotipo {
+    width: 8rem;
+    left: 0;
+    top: 0;
+    position: relative;
+    margin: auto;
+    }
+    
+    .container {
+    width: 50%;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    text-align: center;
+    }
+    
+    .this_button {
+    display: block;
+    background-color: var(--darkbtn-color);
+    color: var(--backgroundmobile-color);
+    padding: 0.5rem;
+    border-radius: 0.5rem;
+    width: 10rem;
+    margin: auto;
+    margin-top: 2rem;
+    border: none;
+    color: var(--backgroundmobile-color);
+    letter-spacing: 0.1rem;
+    font-size: 1rem;
+    text-transform: uppercase;
+    transition: all 0.25s ease-in-out;
+    }
+    
+    .this_button:hover {
+    cursor: pointer;
+    background-color: var(--hoverbotaoescuropraclaro--color);
+    color: black;
+    }
+    }
