--- src/components/homeMobile/homeMobile.css
@@ -1,57 +1,0 @@
-
-@media screen and (min-width: 1000px) {
-    
-}
-
-.container_mobile {
-    flex-direction: column;
-    width: 10rem;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    text-align: center;
-
-}
-
-
-.soft {
-    color: var(--darkbtn-color);
-
-}
-
-h1 span {
-
-    color: #408d86;
-    letter-spacing: 0.2rem;
-}
-
-.btn_mobile {
-    background-color: var(--darkbtn-color);
-    color: var(--backgroundmobile-color);
-    padding: 0.5rem;
-    border-radius: 0.5rem;
-    width: 10rem;
-    margin-top: 2rem;
-    border: none;
-    color: var(--backgroundmobile-color);
-    letter-spacing: 0.1rem;
-    font-size: 1rem;
-    text-transform: uppercase;
-    transition: all 0.25s ease-in-out;
-}
-
-.btn_mobile:hover {
-
-    cursor: pointer;
-    background-color: var(--hoverbotaoescuropraclaro--color);
-    color: black;
-}
-
-@media screen and (min-width: 1000px) {
-
-    .container_mobile{
-        display: none;
-    }
-    
-}
