+++ src/components/navBar/navBar.css
@@ -1,5 +1,3 @@
-
-@media screen and (min-width: 1000px) {
 nav {
     display: block;
     background-color: var(--darkbtn-color);
@@ -7,61 +5,53 @@ nav {
     height: 4rem;
     text-decoration: dotted;
     text-transform: uppercase;
-
-}
-
-nav ul {
+  }
+  
+  nav ul {
     list-style-type: none;
     margin: 0;
     padding: 0;
     text-align: end;
     margin: 1rem;
     margin-right: 7rem;
-
-}
-
-nav li {
+  }
+  
+  nav li {
     display: inline-block;
     margin-right: 2rem;
     color: var(--backgrounddesktop-color);
     text-decoration: none;
     cursor: pointer;
-}
-
-nav li a {
-
+  }
+  
+  nav li a {
     text-decoration: none;
     color: var(--backgrounddesktop-color);
     transition: all 0.25s ease-in-out;
-}
-
-nav li a:hover {
+  }
+  
+  nav li a:hover {
     color: var(--backgroundmobile-color);
-
-}
-
-
-
-.texto_nav:after{
-content: '';
-display: block;
-height: 0.125rem;
-background: var(--backgrounddesktop-color);
-transform: scaleX(0);
-transition: transform .5s;
-}
-.texto_nav:hover:after{
-transform: scaleX(1);
-transition: transform .5s;
-}
-
-
-
-.btnnav_desktop {
-
+  }
+  
+  .texto_nav:after {
+    content: '';
+    display: block;
+    height: 0.125rem;
+    background: var(--backgrounddesktop-color);
+    transform: scaleX(0);
+    transition: transform .5s;
+  }
+  
+  .texto_nav:hover:after {
+    transform: scaleX(1);
+    transition: transform .5s;
+  }
+  
+  .btn {
     height: 2rem;
     border-radius: 0.5rem;
-    width: 5rem;
+    width: 6rem;
     background-color: var(--backgrounddesktop-color);
     border: none;
     font-family: "roboto";
@@ -70,21 +60,16 @@ transition: transform .5s;
     font-weight: 500;
     font-size: 1rem;
     transition: all 0.25s ease-in-out;
-
-
-}
-
-.btnnav_desktop:hover {
-
+  }
+  
+  .btn:hover {
     cursor: pointer;
     background-color: var(--hoverbotaoclaropraescuro--color);
     color: #fafafa;
-}}
-
-@media screen and (max-width: 1000px) {
-
+  }
+  
+  @media screen and (max-width:1000px) {
     nav {
-        display: none;
+      display: none;
     }
-    
-}
+  }
