+++ src/components/navBar/navBar.tsx
@@ -6,14 +6,14 @@ function NavBar() {
   
   <nav>
       <ul>
-        <li className="texto_nav"><a href="">About</a></li>
-        <li className="texto_nav"><a href="">Contact</a></li>
-        <li className="texto_nav"><a href="">Propose</a></li>
+        <li className="texto_nav"><a href="">Sobre</a></li>
+        <li className="texto_nav"><a href="">Contanto</a></li>
+        <li className="texto_nav"><a href="">Propostas</a></li>
         <li>
-          <a href=""><button className="btnnav_desktop">Sign in</button></a>
+          <a href=""><button className="btn">Entrar</button></a>
         </li>
         <li>
-          <a href=""><button className="btnnav_desktop">Sign up</button></a>
+          <a href=""><button className="btn">Registrar</button></a>
         </li>
       </ul>
     </nav>
