+++ src/pages/homePage/homePage.tsx
@@ -1,14 +1,12 @@
 import * as React from "react";
 import NavBar from "../../components/navBar/navBar";
-import { Homedesktop } from "../../components/homeDesktop/homeDesktop";
-import { HomeMobile } from "../../components/homeMobile/homeMobile";
+import { Homedesktop } from "../../components/homeDesktop/homeComponent";
 
 export const Homepage = () => {
   return (
     <>
       <NavBar />
       <Homedesktop />
-      <HomeMobile />
     </>
   );
 };
